Add dataset
This is the dataset of features extracted from objects captured by the horizontal camera.
The "Class" column is the class label. If the value of Class column is OK, the data is some injection. If the value of Class column is NG, the data is some foreign matter (e.g. ample head or vial cap).